requirements.txt
flask
flasgger
pymongo
python-dotenv
